requirements.txt
numpy==2.1.2
GitPython==3.1.43
matplotlib==3.9.2